forgor those

+++ src/r_things.c
@@ -3257,11 +3257,11 @@ static UINT16 W_CheckForSkinMarkerInPwad(UINT16 wadid, UINT16 startlump)
 //sort function for sorting skin names
 static int skinSortFunc(const void *a, const void *b) //tbh i have no clue what the naming conventions for local functions are
 {
-	const skin_t *in1 = &skins[*(const UINT8 *)a];
-	const skin_t *in2 = &skins[*(const UINT8 *)b];
+	const skin_t *in1 = &skins[*(const UINT16 *)a];
+	const skin_t *in2 = &skins[*(const UINT16 *)b];
 	INT32 temp = 0;
-	const UINT8 val_a = *((const UINT8 *)a);
-	const UINT8 val_b = *((const UINT8 *)b);
+	const UINT16 val_a = *((const UINT8 *)a);
+	const UINT16 val_b = *((const UINT8 *)b);
 
 	//return (strcmp(in1->realname, in2->realname) < 0) || (strcmp(in1->realname, in2->realname) ==);
 
@@ -3357,7 +3357,7 @@ static int skinSortFunc(const void *a, const void *b) //tbh i have no clue what 
 void sortSkinGrid(void)
 {
 	//CONS_Printf("Sorting skin list (%d)...\n", cv_skinselectgridsort.value);
-  qs22j(skinsorted, numskins, sizeof(UINT8), skinSortFunc);
+  qs22j(skinsorted, numskins, sizeof(UINT16), skinSortFunc);
 }
 
 //
